refactor(querystring): per-concern split of index.ts
Splits the 1165-LoC `packages/node/querystring/src/index.ts` monolith
(deno_std + Node.js `lib/querystring.js` port) into 5 focused leaf
modules under `src/internal/`, leaving a 68-LoC public entry:

  - internal/tables.ts    (94)  — hexTable, isHexTable, noEscape,
                                  unhexTable lookup tables
  - internal/encode.ts    (101) — encodeStr, qsEscape, ERR_INVALID_URI
  - internal/decode.ts    (82)  — unescapeBuffer, qsUnescape
  - internal/parse.ts     (255) — parse, ParseOptions, charCodes,
                                  addKeyVal
  - internal/stringify.ts (122) — stringify, StringifyOptions,
                                  stringifyPrimitive, encodeStringified*

Dependency graph is a strict DAG (tables ← encode ← stringify,
tables ← decode ← parse, error ← encode); no circular imports.
`src/index.ts` is now pure re-exports + the legacy decode/encode/
escape/unescape aliases + the default-export object that matches
`module.exports` shape for CJS consumers.

Public API surface byte-identical: every name from `node:querystring`
(parse, stringify, escape, unescape, unescapeBuffer, decode, encode,
ERR_INVALID_URI, ParsedUrlQuery type) resolves through the same
symbols. Verified end-to-end: 394/394 byte-identical outputs across
all 256 hex codes + all 128 ASCII escape codes + edge cases;
`tsc --noEmit` clean. No test edits, no package.json#exports change,
no new external runtime deps.

Mirrors the per-concern split shape proven in PR #309
(webgl2-rendering-context.ts → 4 method-group modules) — adapted
for a procedural/util surface (named export groups, no
install*(proto) install functions).
